backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
ccxt==4.4.16
python-binance==1.0.22
celery==5.4.0
redis==5.1.1
python-telegram-bot==21.6
cryptography==43.0.1
python-dotenv==1.0.1
httpx==0.27.2
websockets==13.1
email-validator==2.2.0
aiosmtplib==3.0.1
pyotp==2.9.0
openpyxl==3.1.5
qrcode==7.4.2
pillow==10.4.0
